function for "can person add places here?"
On some games, administration may want to allow people to add places in
rooms that are not strictly "owned" by them. Splitting off this
function allows us to do so.

On The Reach, temporary rooms are maintained by a higher system, but
permission is given to the creator to add places to them.