bug: enactor unfindable made everyone unfindable
If someone set themselves 'unfindable', everybody on 'where' was seen as unfindable, except for wizards and most staffbits.

The problem was that the code was filtering as unfindable whether or not individuals on the 'lwho()' list could see the enactor. A simple swapping of input on '&fil.findable' fixes this issue.