add: preference to show unfindable count per room
On games with a smaller active population count, it became easy to side-step the point of the "unfindable" flag by seeing that a room has one unfindable person, and that person is listed at the bottom of the "where" room result.

In this situation, set the object's '&pref.where.count-unfindable' to '0': Rooms will no longer reveal how many unfindable people are in them.